@@ -6,88 +6,98 @@
   "dependencies": {
     "@babel/preset-react": "7.16.0",
     "@github-docs/frontmatter": "1.3.1",
-    "algoliasearch": "4.11.0",
+    "aether-marketing": "1.27.27",
+    "algoliasearch": "4.14.2",
     "autocomplete.js": "0.37.0",
-    "axios": "0.21.4",
-    "babel-jest": "26.0.0",
-    "babel-preset-gatsby": "2.3.0",
+    "babel-plugin-styled-components": "2.0.7",
     "base-64": "1.0.0",
     "bootstrap": "4.5.0",
     "dotenv": "8.6.0",
     "eslint-plugin-jest": "22.21.0",
-    "gatsby": "4.3.0",
-    "gatsby-cli": "4.3.0",
-    "gatsby-plugin-algolia": "0.24.0",
+    "gatsby": "4.14.1",
+    "gatsby-cli": "4.14.0",
+    "gatsby-plugin-algolia": "1.0.3",
     "gatsby-plugin-env-variables": "2.2.0",
-    "gatsby-plugin-gdpr-cookies": "2.0.8",
-    "gatsby-plugin-google-analytics": "4.3.0",
-    "gatsby-plugin-google-tagmanager": "4.3.0",
-    "gatsby-plugin-manifest": "4.3.0",
+    "gatsby-plugin-gdpr-cookies": "2.0.9",
+    "gatsby-plugin-google-analytics": "4.14.0",
+    "gatsby-plugin-manifest": "4.14.0",
     "gatsby-plugin-meta-redirect": "1.1.1",
-    "gatsby-plugin-newrelic": "2.1.1",
-    "gatsby-plugin-purgecss": "6.1.0",
-    "gatsby-plugin-react-helmet": "5.3.0",
-    "gatsby-plugin-robots-txt": "1.6.14",
-    "gatsby-plugin-sass": "5.3.0",
-    "gatsby-plugin-sharp": "4.3.0",
-    "gatsby-plugin-sitemap": "5.3.0",
+    "gatsby-plugin-newrelic": "2.2.2",
+    "gatsby-plugin-purgecss": "6.1.2",
+    "gatsby-plugin-react-helmet": "5.20.0",
+    "gatsby-plugin-robots-txt": "1.7.1",
+    "gatsby-plugin-sass": "5.20.0",
+    "gatsby-plugin-sharp": "4.14.1",
+    "gatsby-plugin-sitemap": "5.20.0",
     "gatsby-plugin-sri": "1.1.0",
-    "gatsby-plugin-typography": "4.3.0",
-    "gatsby-remark-autolink-headers": "5.3.0",
+    "gatsby-plugin-styled-components": "5.18.0",
+    "gatsby-plugin-typography": "4.17.0",
+    "gatsby-remark-autolink-headers": "5.17.0",
     "gatsby-remark-check-links": "2.1.0",
-    "gatsby-remark-prismjs": "6.3.0",
-    "gatsby-remark-responsive-iframe": "5.3.0",
-    "gatsby-source-filesystem": "4.3.0",
-    "gatsby-transformer-remark": "5.3.0",
-    "gatsby-transformer-sharp": "4.3.0",
+    "gatsby-remark-prismjs": "6.20.0",
+    "gatsby-remark-responsive-iframe": "5.20.0",
+    "gatsby-source-filesystem": "4.14.0",
+    "gatsby-transformer-remark": "5.20.0",
+    "gatsby-transformer-sharp": "4.14.0",
     "glob": "7.1.7",
-    "googleapis": "55.0.0",
-    "identity-obj-proxy": "3.0.0",
+    "googleapis": "128.0.0",
     "jquery": "3.5.1",
     "jquery.scrollto": "2.1.3",
     "markdownlint-cli": "0.19.0",
-    "node-fetch": "2.6.1",
+    "node-fetch": "2.6.7",
     "popper.js": "1.16.1",
     "prettier": "2.5.0",
-    "prismjs": "^1.24.1",
+    "prismjs": "1.28.0",
     "prop-types": "15.7.2",
-    "qs": "6.10.1",
-    "react": "16.14.0",
-    "react-cookie-consent": "4.1.0",
-    "react-dom": "16.14.0",
-    "react-helmet": "5.2.1",
-    "react-instantsearch-dom": "5.7.0",
-    "react-tabs": "3.2.3",
-    "react-test-renderer": "16.14.0",
+    "qs": "6.11.0",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-helmet": "6.1.0",
+    "react-instantsearch": "7.0.1",
     "react-typography": "0.16.20",
-    "s3-deploy": "1.4.0",
-    "sass": "1.44.0",
-    "sendmail": "1.6.1",
-    "shelljs": "0.8.4",
-    "terser": "5.10.0",
+    "sass": "1.32.0",
+    "shelljs": "0.8.5",
+    "styled-components": "5.3.5",
+    "terser": "5.15.1",
     "typescript": "2.9.2",
     "typography": "0.16.21",
+    "unlink": "0.0.2",
     "uuid": "3.3.3",
-    "webpack": "5.64.4"
+    "webpack": "5.89.0"
   },
   "devDependencies": {
     "@babel/eslint-parser": "7.16.3",
+    "@types/jest": "28.1.3",
+    "@types/react-test-renderer": "18.0.0",
+    "babel-jest": "28.1.1",
+    "babel-preset-gatsby": "2.18.0-next.1",
     "eslint": "8.4.1",
-    "jest": "26.0.0"
+    "identity-obj-proxy": "3.0.0",
+    "jest": "28.1.1",
+    "react-test-renderer": "17.0.2"
+  },
+  "optionalDependencies": {
+    "@postman/pmt": "2.0.21"
   },
   "keywords": [
     "gatsby"
   ],
   "license": "MIT",
   "scripts": {
-    "bff": "NODE_ENV=development node bff.js",
-    "build": "npm run bff && GATSBY_NEWRELIC_ENV=production GATSBY_ACTIVE_ENV=development gatsby build",
+    "am": "npm link 'aether-marketing; npm run dev",
+    "bff": "rm -rf bff-data && npm run bff:data && node bff.js",
+    "bff:data": "node -p \"require('./build/bffData.js')()\"",
+    "build": "npm run bff && GATSBY_ACTIVE_ENV=development gatsby build",
     "build:dev": "npm run build",
     "nvmrc": "echo $(node -p -e 'require(\"./package\").engines.node.split(\">=\").join(\"\")') > .nvmrc",
-    "build:prod": "NODE_ENV=production node bff.js && GATSBY_NEWRELIC_ENV=production GATSBY_ACTIVE_ENV=production gatsby build",
+    "package": "cp -rf package/* node_modules/",
+    "postinstall": "npm run package",
+    "build:prod": "NODE_ENV=production npm run bff && GATSBY_ACTIVE_ENV=production gatsby build",
     "clean": "gatsby clean",
-    "dev": "npm run bff && gatsby develop",
+    "dev": "node lint-front.js && npm run bff && gatsby develop",
     "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
+    "reset": "rm -rf package-lock.json; rm -rf node_modules; npm i; ./node_modules/.bin/gatsby clean;",
+    "sanitize-branch": "node sanitize-branch.js",
     "start": "npm run develop",
     "serve": "gatsby serve",
     "test": "npm run bff && GATSBY_ACTIVE_ENV=development && jest",
@@ -105,7 +115,7 @@
     "url": "https://github.com/postmanlabs/postman-docs/issues"
   },
   "engines": {
-    "node": "14.15.0"
+    "node": "14.17.1"
   },
   "allow": {
     "pmTech": [
