@@ -7,11 +7,9 @@ const { v4: uuidv4 } = require('uuid');
 const frontmatter = require('@github-docs/frontmatter');
 const redirects = require('./redirects.json');
 // const HeaderJson = require('./src/components/Header/Header.data.json');
-const FooterJson = require('./src/components/Footer/Footer.data.json');
 const { execSync } = require("child_process")
 const ignorePaths = [];
 
-// const { google } = require('googleapis');
 
 exports.onCreateNode = ({ node, getNode, actions }) => {
   const { createNodeField } = actions;
@@ -22,7 +20,8 @@ exports.onCreateNode = ({ node, getNode, actions }) => {
       name: 'slug',
       value: slug,
     });
-    /* Returns the latest commit log for a specific doc file (View Doc.jsx for query) */
+    /* Used on doc.jsx / View GraphQL query
+    /* Returns the latest commit log for a specific doc file */
     const lastModifiedDate = execSync(
       `git log -1 --pretty='%ad' --date=format:'%Y/%m/%d' ${node.fileAbsolutePath}`
     ).toString()
@@ -31,7 +30,22 @@ exports.onCreateNode = ({ node, getNode, actions }) => {
       name: "lastModifiedDate",
       value: lastModifiedDate,
     })
+    /* Returns the last modified time (SEO)  */
+    const lastModifiedTime = execSync(
+      `git log -1 --pretty="format:%cI" ${node.fileAbsolutePath}`
+    ).toString();
+    actions.createNodeField({
+      node,
+      name: "lastModifiedTime",
+      value: lastModifiedTime,
+    })
   }
+  // display console error if algolia indexed frontmatter length exceeds 9999 bytes on build
+  // if (node.internal.type === 'frontmatterLength') {
+  //   if(JSON.parse(node.internal.content).value > 9999) {
+  //     console.error("IMPORTANT: Frontmatter has too many characters");
+  //   }
+  // }
 };
 
 exports.createPages = async ({ graphql, actions }) => {
@@ -80,7 +94,7 @@ exports.createPages = async ({ graphql, actions }) => {
   });
 };
 
-/* Create Header and Footer
+/* Create Header
 /************************************************************************ */
 exports.sourceNodes = async ({
   actions,
@@ -108,14 +122,18 @@ exports.sourceNodes = async ({
     const output = { ...data, ...nodeMeta };
     return output;
   };
-
   const { createNode } = actions;
-
+  let mdFrontmatterCharacterCount = [];
+  let excerptCharacterCount = [];
   const getDirectories = (src) => glob.sync(`${src}/**/*`);
   const paths = getDirectories('./src/pages/docs')
     .filter((val) => val.slice(-3) === '.md')
     .map((val) => {
       const { data } = frontmatter(fs.readFileSync(val));
+      mdFrontmatterCharacterCount.push(data.title, data.search_keyword);
+      excerptCharacterCount.push(data.excerpt)
+      // const algoliaLength = JSON.stringify(val[data]);
+      // h = algoliaLength;
       const order = data.order || 200;
       return [val, order];
     })
@@ -130,20 +148,26 @@ exports.sourceNodes = async ({
     .filter((val) => !ignorePaths.includes(val));
 
   const output = {};
-
+  
   paths.forEach((val) => {
     let split = val.split('/');
     split = split.filter((url) => url !== '');
 
     let current = output;
     split.forEach((part) => {
-      current[part] = current[part] || {};
+      current[part] = current[part] || {}; 
       current = current[part];
     });
     current.url = `/${split.join('/')}/`;
   });
-
+  mdFrontmatterCharacterCount = JSON.stringify(mdFrontmatterCharacterCount);
+  // enable console.logs to view frontmatter object in terminal 'npm run dev'
+  // console.log(mdFrontmatterCharacterCount, 'List of all frontmatter objects from md files')
+  mdFrontmatterCharacterCount = mdFrontmatterCharacterCount.length;
+  // console.log('Length of all frontmatter from md files', mdFrontmatterCharacterCount)
+  excerptCharacterCount = JSON.stringify(excerptCharacterCount);
+  excerptCharacterCount = excerptCharacterCount.length;
+  
+  createNode(prepareNode(mdFrontmatterCharacterCount, 'frontmatterLength'));
   createNode(prepareNode(output.docs, 'leftNavLinks'));
-  // createNode(prepareNode(HeaderJson, 'headerLinks'));
-  createNode(prepareNode(FooterJson, 'FooterLinks'));
 };
