@@ -1,5 +1,5 @@
 const babelOptions = {
-  presets: ['babel-preset-gatsby'],
-};
+  presets: ["babel-preset-gatsby"],
+}
 
-module.exports = require('babel-jest').createTransformer(babelOptions);
+module.exports = require("babel-jest").createTransformer(babelOptions);
