@@ -1,16 +1,18 @@
 module.exports = {
   transform: {
-    '^.+\\.jsx?$': '<rootDir>/jest-preprocess.js',
+    "^.+\\.jsx?$": `<rootDir>/jest-preprocess.js`,
   },
   moduleNameMapper: {
-    '.+\\.(css|styl|less|sass|scss)$': 'identity-obj-proxy',
-    '.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>/__mocks__/file-mock.js',
+    ".+\\.(css|styl|less|sass|scss)$": `identity-obj-proxy`,
+    ".+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": `<rootDir>/__mocks__/file-mock.js`
   },
-  testPathIgnorePatterns: ['node_modules', '.cache', 'public'],
-  transformIgnorePatterns: ['node_modules/(?!(gatsby)/)'],
+  testPathIgnorePatterns: [`node_modules`, `\\.cache`, `<rootDir>.*/public`],
+  transformIgnorePatterns: [`node_modules/(?!(gatsby|gatsby-script)/)`],
   globals: {
-    __PATH_PREFIX__: '',
+    __PATH_PREFIX__: ``,
   },
-  testURL: 'http://localhost',
-  setupFiles: ['<rootDir>/loadershim.js'],
-};
+  testEnvironmentOptions: {
+    url: `http://localhost`
+  },
+  setupFiles: [`<rootDir>/loadershim.js`],
+}
