@@ -1,10 +1,13 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import Footer from './Footer';
+import Theme from '../../../styles/theme.jsx'
 
+describe('Footer', () => {
 it('renders correctly', () => {
   const tree = renderer
-    .create(<Footer />)
+    .create(<Theme><Footer /></Theme>)
     .toJSON();
   expect(tree).toMatchSnapshot();
+  });
 });
