@@ -1,9 +1,16 @@
-import React from 'react';
-import './Header.scss';
+import React, { useEffect, useState } from 'react';
 import { trackCustomEvent } from 'gatsby-plugin-google-analytics';
 import Dropdown from './Dropdown';
 import $ from 'jquery';
-import postmanLogo from '../../images/postman-logo-icon.svg';
+import {PrimaryNavbarV6, SecondaryNavbarV6, NavStyles, DropdownStyles, DropdownStylesSecond, CTAButton} from './HeaderStyles.jsx' ;
+import { SearchWrapperStyling } from '../Search/searchStyles.jsx';
+import { Paragraph } from 'aether-marketing';
+// prod nav data
+import navbarData from '../../../bff-data/navbar.json';
+import navtopicsdropdownData from '../../../bff-data/navtopicsdropdown.json';
+// fallback nav data
+import navbarDataLocal from '../../../build/navbarDev.json';
+import navtopicsdropdownDataLocal from '../../../build/navtopicsdropdownDev.json';
 
 // Get Cookie for Sign In toggler
 const getCookie = (a) => {
@@ -20,7 +27,24 @@ const LoginCheck = (props) => {
 
   if (!hidden) {
     return (
-      <>
+      <CTAButton>
+        <a
+          href={`https://go.postman${beta}.co/build`}
+          className='button__sign-in pingdom-transactional-check__sign-in-button'
+          style={{ padding: '4px 12px 4px 12px' }}
+          onClick={() => {
+            trackCustomEvent({
+              // string - required - The object that was interacted with (e.g.video)
+              category: 'lc-top-nav',
+              // string - required - Type of interaction (e.g. 'play')
+              action: 'Click',
+              // string - optional - Useful for categorizing events (e.g. 'Spring Campaign')
+              label: 'contact-sales-button-clicked',
+            });
+          }}
+        >
+         Contact Sales
+        </a>
         <a
           href={`https://go.postman${beta}.co/build`}
           className={
@@ -60,36 +84,47 @@ const LoginCheck = (props) => {
         >
           Launch Postman
         </a>
-      </>
+      </CTAButton>
     );
   }
   return <></>;
 };
-class Header extends React.Component {
-  constructor(props) {
-    super(props);
 
-    this.state = {
-      beta: '',
-      cookie: '',
-      hidden: true,
-    };
-  }
+const Header = (props) => {
 
-  componentDidMount() {
+  const [beta, setBeta] = useState('');
+  const [cookie, setCookie] = useState('');
+  const [hidden, setHidden] = useState(true);
+  const [data, setData] = useState(navbarData);
+  const [dataDropdown, setDataDropdown] = useState(navtopicsdropdownData);
+  const [visibleHelloBar] = useState();
+
+  useEffect(() => {
     const cookie = getCookie('getpostmanlogin');
     const beta = window.location.host.includes('postman-beta') ? '-beta' : '';
 
-    this.setState({
-      cookie,
-      beta,
-    });
+    setCookie(cookie);
+    setBeta(beta);
+    const navbarKeys = ['items', 'media', 'type'];
+    const navtopicsdropdownKeys = ['items', 'type'];
+    
+    if (navbarKeys.every(key => Object.keys(navbarData).includes(key))) {
+      setData(navbarData)
+    } else {
+      setData(navbarDataLocal)
+    }
 
-    /* eslint-disable react/prop-types */
-    const { waitBeforeShow } = this.props;
-    /* eslint-enable react/prop-types */
+
+    if (navtopicsdropdownKeys.every(key => Object.keys(navtopicsdropdownData).includes(key))) {
+      setDataDropdown(navtopicsdropdownData)
+    } else {
+      setDataDropdown(navtopicsdropdownDataLocal)
+    }
+
+    const { waitBeforeShow } = props;
+
     setTimeout(() => {
-      this.setState({ hidden: false });
+      setHidden(false)
     }, waitBeforeShow);
 
     /* Applies styling for sticky nav */
@@ -98,32 +133,33 @@ class Header extends React.Component {
       $('.nav-primary').toggleClass('activeMenu');
       $('.nav-secondary').toggleClass('activeMenu');
     });
-    // Dropdown Slideup Animation
+    // Toggles Dropdown Menu and Fade Animation
     function showBsDropdown() {
       $(this)
         .find('.dropdown-menu')
         .first()
         .stop(true, true)
-        .slideDown(225);
+        .fadeToggle(250)
       $(this)
         .find('.arrow-icon')
         .addClass('show');
     }
     $('.dropdown').on('show.bs.dropdown', showBsDropdown);
-    // Dropdown Slidedown Animation
+
+    // Unbinds Dropdown Menu and Fade Animation
     function hideBsDropdown() {
       $(this)
         .find('.dropdown-menu')
         .stop(true, true)
-        .slideUp(225);
+        .fadeToggle(250)
       $(this)
         .find('.arrow-icon')
         .removeClass('show');
     }
     $('.dropdown').on('hide.bs.dropdown', hideBsDropdown);
-  }
+  }, [])
 
-  showTargetElement = () => {
+ const showTargetElement = () => {
     // Show Sign In Button if user is not logged in (mobile)
     const cookie = getCookie('getpostmanlogin');
     const signInButton = document.querySelector('.mobile-sign-in');
@@ -141,7 +177,7 @@ class Header extends React.Component {
       // Add lock CSS to body to disable scroll
       body.classList.add('lock');
       // Flip up dropdown icon
-      icon1.classList.add('open');
+      icon1.classList.remove('open');
     }
     // Hellobar
     const messageBarAlertTop = document.getElementById(
@@ -154,7 +190,7 @@ class Header extends React.Component {
     }
   }
 
-  hideTargetElement = () => {
+  const hideTargetElement = () => {
     // Hide Sign In Button if user is not logged in (mobile)
     const signInButton = document.querySelector('.mobile-sign-in');
     const cookie = getCookie('getpostmanlogin');
@@ -168,10 +204,13 @@ class Header extends React.Component {
     const icon1 = document.getElementById('icon-wrap-one');
     // Mobile Menu is not active ?
     if (toggler === 'false') {
+      // addlock CSS to body to enable scroll
+      body.classList.add('lock');
+      // Flip down dropdown icon
+      icon1.classList.add('open');
+    } else {
       // Remove lock CSS to body to disable scroll
       body.classList.remove('lock');
-      // Flip down dropdown icon
-      icon1.classList.remove('open');
     }
     const icon2 = document.getElementById('navbar-chevron-icons');
     const togglerSecondary = document
@@ -182,44 +221,34 @@ class Header extends React.Component {
     }
   }
 
-  showTargetElementLC = () => {
+  const toggleSecondaryNavMenu = () => {
     // LC Mobile Icon Transition
     const togglerSecondary = document
       .getElementById('secondaryNav')
       .getAttribute('aria-expanded');
     const toggleChevron = document.getElementById('navbar-chevron-icons');
-    if (togglerSecondary === 'true') {
-      toggleChevron.classList.add('open');
-    }
-  }
-
-  hideTargetElementLC = () => {
-    const toggleChevron = document.getElementById('navbar-chevron-icons');
-    const togglerSecondary = document
-      .getElementById('secondaryNav')
-      .getAttribute('aria-expanded');
     if (togglerSecondary === 'false') {
+      toggleChevron.classList.add('open');
+    } else {
       toggleChevron.classList.remove('open');
     }
   }
+
   /* eslint-enable class-methods-use-this */
 
-  render() {
-    const {
-      beta, visibleHelloBar, cookie, hidden,
-    } = this.state;
-    return (
-      <>
-        <nav className="navbar-v6 navbar navbar-expand-lg navbar-light bg-light nav-primary">
+  return (
+    <>
+      <PrimaryNavbarV6 className="navbar-v6 ">
+        <NavStyles className="navbar navbar-expand-lg navbar-light nav-primary ">
           <a className="navbar-brand" href="https://www.postman.com">
             <div className="navbar-logo-container">
-              <img src={postmanLogo} alt="Postman" width="32" height="32" />
+              <img src="https://voyager.postman.com/logo/postman-logo-icon-orange.svg" alt="Postman" width="32" height="32" />
             </div>
           </a>
           <button
             onClick={() => {
-              this.showTargetElement();
-              this.hideTargetElement();
+              showTargetElement();
+              hideTargetElement();
             }}
             id="globalNav"
             className="mobile-sign-in navbar-toggler"
@@ -246,288 +275,86 @@ class Header extends React.Component {
           </button>
           <div
             id="navbarSupportedContent"
-            className={`collapse navbar-collapse${!visibleHelloBar ? ' noBar' : ''
-              }`}
+            className={`collapse navbar-collapse ${!visibleHelloBar ? 'noBar' : ''}`}
           >
+            {/* Primary Navbar */}
             <ul className="navbar-nav mr-auto">
-              <li className="nav-item dropdown">
-                <a
-                  className="nav-link dropdown-toggle"
-                  href="##"
-                  id="navbarDropdownMenuLink"
-                  data-toggle="dropdown"
-                  aria-expanded="false"
-                >
-                  Product
-                  <svg
-                    className="arrow-icon"
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="12"
-                    height="12"
-                    viewBox="0 0 12 12"
-                    fill="#6b6b6b"
-                  >
-                    <g>
-                      <path d="M10.375,3.219,6,6.719l-4.375-3.5A1,1,0,1,0,.375,4.781l5,4a1,1,0,0,0,1.25,0l5-4a1,1,0,0,0-1.25-1.562Z" />
-                    </g>
-                  </svg>
-                </a>
-                <div
-                  className="dropdown-menu"
-                  aria-labelledby="navbarDropdownMenuLink"
-                >
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/product/what-is-postman/"
-                  >
-                    What is Postman?
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/product/api-repository/"
-                  >
-                    API repository
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/product/tools/"
-                  >
-                    Tools
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/product/intelligence/"
-                  >
-                    Intelligence
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/product/workspaces/"
-                  >
-                    Workspaces
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/product/integrations/"
-                  >
-                    Integrations
-                  </a>
-                  <a
-                    className="dropdown-item app-cta"
-                    href="https://www.postman.com/downloads/"
-                  >
-                    Get started free →
-                  </a>
-                </div>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="https://www.postman.com/pricing/">
-                  Pricing
-                </a>
-              </li>
-              <li className="nav-item dropdown">
-                <a
-                  className="nav-link dropdown-toggle"
-                  href="##"
-                  id="navbarDropdownMenuLink"
-                  data-toggle="dropdown"
-                  aria-expanded="false"
-                >
-                  Enterprise
-                  <svg
-                    className="arrow-icon"
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="12"
-                    height="12"
-                    viewBox="0 0 12 12"
-                    fill="#6b6b6b"
-                  >
-                    <g>
-                      <path d="M10.375,3.219,6,6.719l-4.375-3.5A1,1,0,1,0,.375,4.781l5,4a1,1,0,0,0,1.25,0l5-4a1,1,0,0,0-1.25-1.562Z" />
-                    </g>
-                  </svg>
-                </a>
-                <div
-                  className="dropdown-menu"
-                  aria-labelledby="navbarDropdownMenuLink"
-                >
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/postman-enterprise/"
-                  >
-                    Postman Enterprise
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/case-studies/"
-                  >
-                    Enterprise case studies
-                  </a>
-                  <a
-                    className="dropdown-item"
-                    href="https://www.postman.com/company/contact-sales/"
-                  >
-                    Contact sales
-                  </a>
-                </div>
-              </li>
-              <li className="nav-item dropdown">
-                <a
-                  className="nav-link dropdown-toggle"
-                  href="##"
-                  id="navbarDropdownMenuLink"
-                  data-toggle="dropdown"
-                  aria-expanded="false"
-                >
-                  Resources and support
-                  <svg
-                    className="arrow-icon"
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="12"
-                    height="12"
-                    viewBox="0 0 12 12"
-                    fill="#6b6b6b"
-                  >
-                    <g>
-                      <path d="M10.375,3.219,6,6.719l-4.375-3.5A1,1,0,1,0,.375,4.781l5,4a1,1,0,0,0,1.25,0l5-4a1,1,0,0,0-1.25-1.562Z" />
-                    </g>
-                  </svg>
-                </a>
-                <div
-                  className="dropdown-menu"
-                  aria-labelledby="navbarDropdownMenuLink"
-                >
-                  <div className="row dropdown-col-menu">
-                    <div className="col-sm-6 col-md-4 dropdown-col">
-                      <h6 className="dropdown-header">Learning</h6>
-                      <a
-                        className="dropdown-item"
-                        href="/docs/getting-started/introduction/"
+            {data.items.map((item) => (
+                item.dropdown && item.dropdown && (
+                  <li className="nav-item dropdown" key={item.title}>
+                    <a
+                      className="nav-link dropdown-toggle"
+                      href="##"
+                      id="navbarDropdownMenuLink"
+                      data-toggle="dropdown"
+                      aria-expanded="false"
+                      key={item.title}
+                    >
+                      {item.title}
+                      <svg
+                        className="arrow-icon"
+                        xmlns="http://www.w3.org/2000/svg"
+                        width="12"
+                        height="12"
+                        viewBox="0 0 12 12"
+                        fill="#6b6b6b"
                       >
-                        Docs
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/webinars/"
-                      >
-                        Webinars
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/events/breaking-changes/"
-                      >
-                        Breaking Changes show
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/events/postman-space-camp/"
-                      >
-                        Postman Space Camp
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/case-studies/"
-                      >
-                        Case studies
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/state-of-api/"
-                      >
-                        State of the API
-                      </a>
-                    </div>
-                    <div className="col-sm-6 col-md-4 dropdown-col">
-                      <h6 className="dropdown-header">Community and events</h6>
-                      <a
-                        className="dropdown-item"
-                        href="https://blog.postman.com/"
-                      >
-                        Blog
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/community/"
-                      >
-                        Community
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/company/student-program/"
-                      >
-                        Student program
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/company/student-program/student-summit/"
-                      >
-                        Student Summit
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/postman-galaxy/"
-                      >
-                        Postman Galaxy
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/company/nonprofit/"
-                      >
-                        Postman for Nonprofits
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://store.postman.com/"
-                        target="_blank"
-                        rel="noreferrer"
-                      >
-                        Postman swag
-                      </a>
-                    </div>
-                    <div className="col-sm-6 col-md-4 dropdown-col">
-                      <h6 className="dropdown-header">Support</h6>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/support/"
-                      >
-                        Support Center
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/support/resellers-support/"
-                      >
-                        Reseller support
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://status.postman.com/"
-                        target="_blank"
-                        rel="noreferrer"
-                      >
-                        Postman status
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/downloads/release-notes/"
-                      >
-                        Release notes
-                      </a>
-                      <a
-                        className="dropdown-item"
-                        href="https://www.postman.com/company/contact-us/"
-                      >
-                        Contact us
-                      </a>
-                    </div>
-                  </div>
-                </div>
-              </li>
-              <li className="nav-item">
-                <a className="nav-link" href="https://www.postman.com/explore">
-                  Explore
-                </a>
-              </li>
-            </ul>
+                        <g>
+                          <path d="M10.375,3.219,6,6.719l-4.375-3.5A1,1,0,1,0,.375,4.781l5,4a1,1,0,0,0,1.25,0l5-4a1,1,0,0,0-1.25-1.562Z" />
+                        </g>
+                      </svg>
+                    </a>
+                      <DropdownStyles
+                          className="dropdown-menu dropdown-primary"
+                          aria-labelledby="navbarDropdownMenuLink"
+                        >
+                          { item.columns && item.columns &&
+                          <div className="row dropdown-col-menu">
+                            { item.columns.map((col) => (
+                              <div
+                              className={
+                                item.isWidthShort
+                                  ? 'col-sm-6 col-md-6 dropdown-col'
+                                  : 'col-sm-6 col-md-4 dropdown-col'
+                              } key={col.title}>
+                                <h6 className="dropdown-header">{col.title}</h6>
+                                {col.subItemsCol.map((link) => (
+                                  <a
+                                    className={`${link.link ? `${link.link}` : ''} dropdown-item`}
+                                    href={link.url}
+                                    key={link.title}
+                                  >
+                                    {link.title}
+                                  </a>
+                                ))}
+                              </div>
+                            ))}
+                          </div> || item.subItems.map((single) => (
+                          <a
+                            className={`${single.link ? 'app-cta' : ''} dropdown-item`}
+                            href={single.url}
+                            key={single.title}
+                          >
+                            {single.title}
+                          </a>
+                        ))}
+                        </DropdownStyles>
+                  </li>
+                ) || (
+                  <li className="nav-item" key={item.title}>
+                    <a
+                      className="nav-link"
+                      href={item.url}
+                      key={item.title}>
+                      {item.title}
+                    </a>
+                  </li>
+                )
+              )
+            )
+          }
+          </ul>
+          {/* Login Check */}
             <div className="form-inline my-2 my-lg-0">
               <LoginCheck
                 hidden={hidden}
@@ -538,21 +365,56 @@ class Header extends React.Component {
               />
             </div>
           </div>
-        </nav>
-        <nav className="navbar-v6 navbar navbar-expand-lg navbar-light bg-light nav-secondary blurred-container">
-          <a
-            className="navbar-brand"
-            href="/docs/getting-started/introduction/"
-          >
-            <span id="learning-center-home-link" className="nav-link uber-nav">
-              Learning Center
-              <span className="sr-only">(current)</span>
-            </span>
-          </a>
+        </NavStyles>
+      </PrimaryNavbarV6>
+      <SecondaryNavbarV6 className="navbar-v6 sticky ">
+        <NavStyles className="navbar navbar-expand-lg navbar-light nav-secondary blurred-container">
+            <DropdownStylesSecond className="dropdown position-static">
+              <a
+                className="nav-link navbar-brand"
+                href="/"
+                id="navbarDropdownMenuLink"
+                data-toggle="dropdown"
+                aria-expanded="false"
+              >
+                Learning Center
+                <svg
+                  className="arrow-icon"
+                  xmlns="http://www.w3.org/2000/svg"
+                  width="12"
+                  height="12"
+                  viewBox="0 0 12 12"
+                  fill="#212121"
+                >
+                  <g>
+                    <path d="M10.375,3.219,6,6.719l-4.375-3.5A1,1,0,1,0,.375,4.781l5,4a1,1,0,0,0,1.25,0l5-4a1,1,0,0,0-1.25-1.562Z" />
+                  </g>
+                </svg>
+              </a>
+              <div className="dropdown-menu lc-iconic">
+                <ul>
+                  {dataDropdown.items.map((item) => (
+                    <li key={item.title}>
+                    <a href={item.url} className="dropdown-item mb-3">
+                      <div className="row">
+                        <div className="col-1 lc-icon">
+                          <img className="d-block mx-auto" src={item.img} height="40px"/>
+                        </div>
+                        <div className="col-11">
+                          <Paragraph className="strong mb-0" >{item.title} <svg className="arrow-icon" xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 12 12" fill="#212121"><g><path d="M10.375,3.219,6,6.719l-4.375-3.5A1,1,0,1,0,.375,4.781l5,4a1,1,0,0,0,1.25,0l5-4a1,1,0,0,0-1.25-1.562Z"></path></g></svg></Paragraph>
+                          <Paragraph className="dropdown-item-text-wrap small" >{item.body}</Paragraph>
+                        </div>
+                      </div>
+                    </a>
+                  </li>
+                  ))}
+
+                </ul>
+              </div>
+            </DropdownStylesSecond>
           <button
             onClick={() => {
-              this.showTargetElementLC();
-              this.hideTargetElementLC();
+              toggleSecondaryNavMenu();
             }}
             id="secondaryNav"
             className="mobile-sign-in navbar-toggler"
@@ -579,34 +441,8 @@ class Header extends React.Component {
             className="collapse navbar-collapse"
             id="navbarSupportedContentBottom"
           >
-            <ul className="property-context-menu navbar-nav ml-auto">
-              <li className="nav-item">
-                <a
-                  className="nav-link uber-nav"
-                  href="/docs/getting-started/introduction/"
-                >
-                  Docs
-                </a>
-              </li>
-              <li className="nav-item">
-                <a
-                  className="nav-link uber-nav"
-                  href="/docs/administration/managing-your-team/managing-your-team/"
-                >
-                  Admin
-                </a>
-              </li>
-              <li className="nav-item">
-                <a
-                  className="nav-link uber-nav mr-3"
-                  href="/docs/developer/resources-intro/"
-                >
-                  Developer
-                </a>
-              </li>
-            </ul>
-            {/* Aloglia Widgets */}
-            <div className="form-inline header__search">
+            {/* Algolia Widgets */}
+            <SearchWrapperStyling className="form-inline header__search  ml-auto">
               <svg
                 className="nav-search__icon"
                 width="16"
@@ -621,14 +457,14 @@ class Header extends React.Component {
                   d="M9.87147 9.16437C10.5768 8.30243 11 7.20063 11 6C11 3.23858 8.76142 1 6 1C3.23858 1 1 3.23858 1 6C1 8.76142 3.23858 11 6 11C7.20063 11 8.30243 10.5768 9.16437 9.87147L9.89648 10.6036L9.64648 10.8536L13.5758 14.7829C13.8101 15.0172 14.19 15.0172 14.4243 14.7829L14.7829 14.4243C15.0172 14.19 15.0172 13.8101 14.7829 13.5758L10.8536 9.64648L10.6036 9.89648L9.87147 9.16437ZM6 10C8.20914 10 10 8.20914 10 6C10 3.79086 8.20914 2 6 2C3.79086 2 2 3.79086 2 6C2 8.20914 3.79086 10 6 10Z"
                 />
               </svg>
-            
+
               <Dropdown />
-            </div>
+            </SearchWrapperStyling>
           </div>
-        </nav>
-      </>
-    );
-  }
-}
+        </NavStyles>
+      </SecondaryNavbarV6>
+    </>
+  );
+;}
 
 export default Header;
