@@ -1,34 +1,61 @@
 import React from 'react';
 import { OutboundLink } from 'gatsby-plugin-google-analytics';
-import './Cards.scss';
-import './Buttons';
+import styled from 'styled-components';
+import { BaseLink } from 'aether-marketing'
+
+const CardWrapper = styled.div`
+  margin-bottom: 32px;
+  text-align: left !important;
+  .landing-card__top {
+    background: transparent;
+    margin: 0;
+  }
+  .landing-card__image {
+    margin-bottom: 16px;
+    width: 50px;
+    & img {
+      max-width: 100%;
+    }
+  }
+  .landing-card__content {
+    &-description {
+      margin-bottom: 25px;
+      font-size: 16px;
+    }
+    a {
+      margin-bottom: 48px;
+    }
+  }
+`
 
 export const LandingCard = ({
   title, description, link, icon, cta,
 }) => (
-  <div className="landing-card h-100">
+  <CardWrapper className="landing-card h-100">
     <div className="landing-card__top">
       <div className="landing-card__image">
-        <img src={icon} alt="landing-card icon" />
+        <img src={icon} alt={title} aria-hidden="true"/>
       </div>
     </div>
     <div className="landing-card__content text-left">
       <h3 className="landing-card__content-title h4">{title}</h3>
       <p className="landing-card__content-description">{description}</p>
-      <a href={link} className="link-style">
+      <BaseLink 
+        src={link} 
+        linkType="arrowLink"
+        target="same-tab"
+      >
         {cta}
-        {' '}
-        →
-      </a>
+      </BaseLink>
     </div>
-  </div>
+  </CardWrapper>
 );
 
 // Landing Card with containing an outbound link
 export const LandingCardWithOutboundLink = ({
   title, description, link, icon, cta,
 }) => (
-  <div className="landing-card h-100">
+  <CardWrapper className="landing-card h-100">
     <div className="landing-card__top">
       <div className="landing-card__image">
         <img src={icon} alt="landing-card icon" />
@@ -39,7 +66,7 @@ export const LandingCardWithOutboundLink = ({
       <p className="landing-card__content-description">{description}</p>
       <OutboundLink href={link} className="btn btn__primary">{cta}</OutboundLink>
     </div>
-  </div>
+  </CardWrapper>
 );
 
 export const SmallCard = ({
