@@ -1,5 +1,28 @@
 import React, { Component } from 'react';
+import { BaseButton } from 'aether-marketing';
+import styled from 'styled-components';
 
+const SvgWrapper = styled.i`
+  // ICONS
+  // class starting with "pm-icon-" get styled by _pm-icons.scss file.
+  // class .glyphicon gets styled by bootstrap.css file.
+
+
+  // gets rid of weird background bug with our icons. We need to redo these.
+  background: 0 0;
+  
+  svg.button-icon--left {
+    margin-right: 7px;
+    overflow: hidden;
+    vertical-align: middle;
+  }
+`
+
+
+const EditbuttonStyle = styled.span`
+  font-family: inter;
+  font-size: 14px;
+`
 class EditDoc extends Component {
   constructor(props) {
     super(props);
@@ -15,14 +38,20 @@ class EditDoc extends Component {
     });
   }
 
+
   render() {
-    const { className } = this.props;
-    const classes = className ? `${className}` : '';
     const { pathRoute } = this.state;
     return (
-      <a id="GTM-LC-id" className={classes} href={`https://github.com/postmanlabs/postman-docs/blob/develop/src/pages${pathRoute}.md`} target="_blank" rel="noopener nofollow">
-        <i>
-          <svg className="button-icon--left" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16">
+      <BaseButton 
+        buttonType="secondary"
+        as="a"
+        id="GTM-LC-id" 
+        src={`https://github.com/postmanlabs/postman-docs/blob/develop/src/pages${pathRoute}.md`} 
+        target="new-tab-external-nofollow"
+        className="pmt-edit-btn"
+        >
+        <SvgWrapper >
+          <svg className="button-icon--left" transform="translate(0, -1)" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16">
             <title>logo-github</title>
             <g fill="#212121"><path fillRule="evenodd" clipRule="evenodd" fill="#212121" d="M8,0.2c-4.4,0-8,3.6-8,8c0,3.5,2.3,6.5,5.5,7.6 C5.9,15.9,6,15.6,6,15.4c0-0.2,0-0.7,0-1.4C3.8,14.5,3.3,13,3.3,13c-0.4-0.9-0.9-1.2-0.9-1.2c-0.7-0.5,0.1-0.5,0.1-0.5 c0.8,0.1,1.2,0.8,1.2,0.8C4.4,13.4,5.6,13,6,12.8c0.1-0.5,0.3-0.9,0.5-1.1c-1.8-0.2-3.6-0.9-3.6-4c0-0.9,0.3-1.6,0.8-2.1 c-0.1-0.2-0.4-1,0.1-2.1c0,0,0.7-0.2,2.2,0.8c0.6-0.2,1.3-0.3,2-0.3c0.7,0,1.4,0.1,2,0.3c1.5-1,2.2-0.8,2.2-0.8 c0.4,1.1,0.2,1.9,0.1,2.1c0.5,0.6,0.8,1.3,0.8,2.1c0,3.1-1.9,3.7-3.7,3.9C9.7,12,10,12.5,10,13.2c0,1.1,0,1.9,0,2.2 c0,0.2,0.1,0.5,0.6,0.4c3.2-1.1,5.5-4.1,5.5-7.6C16,3.8,12.4,0.2,8,0.2z" /></g>
           </svg>
@@ -40,9 +69,11 @@ class EditDoc extends Component {
               d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.013 8.013 0 0 0 16 8c0-4.42-3.58-8-8-8z"
             />
           </svg> */}
-        </i>
-        Edit this doc
-      </a>
+        </SvgWrapper>
+        <EditbuttonStyle>
+          Edit This Doc
+        </EditbuttonStyle>
+      </BaseButton>
     );
   }
 }
