@@ -1,69 +1,57 @@
 import React from 'react';
-import { OutboundLink } from 'gatsby-plugin-google-analytics';
-import { v4 as uuidv4 } from 'uuid';
+import styled from 'styled-components';
+import '../../styles/config/normalize.css';
 import Layout from '../components/layout';
 import SEO from '../components/seo';
-import '../../styles/config/_pm-icons.css';
-import upcomingEvents from '../../bff-data/events.json';
-import HeroImage from '../assets/refresh-hero-image.svg';
-// import TeamCollaborationImage from '../assets/team-collaboration-postmanaut-pic.png';
-
-import LearnImage from '../assets/icons/spaceship.svg';
-import ThirtyDaysImage from '../assets/icons/trophy.svg';
-import SpaceCampImage from '../assets/icons/camp.svg';
-import NetworkImage from '../assets/icons/hotspot.svg';
-import TestImage from '../assets/icons/flask.svg';
-import VideoImage from '../assets/icons/movie.svg';
-import SupportImage from '../assets/icons/support.svg';
-import BugImage from '../assets/icons/bug.svg';
-import CommunityImage from '../assets/icons/community.svg';
-import SolutionImage from '../assets/icons/solution.svg';
-
 import { LandingCard } from '../components/MarketingPages/Cards';
-import '../../styles/config/normalize.css';
-import '../components/MarketingPages/Buttons.scss';
-import './index.scss';
+import '../../styles/config/_pm-icons.css';
+import { BaseLink, BaseButton, SectionStyles, VideoComponent } from 'aether-marketing';
 
+const HeroWrapper = styled.section`
+  background-color: rgba(173, 205, 251, .2);
+  padding: 48px 80px;
+    @media (max-width: 991px) {
+        padding: 40px !important;
+      }
+    .hero-image {
+        margin: 0px;
+    }
+    .img-frame {
+        border-radius: ${(props) => props.theme.borderRadius.medium};
+        border: 8px solid ${(props) => props.theme.colors.grey_20};// $grey_20
+        box-shadow: 8px 8px 0 rgba(0, 0, 0, 0.32);
+    }
+`
 
-const heroBackground = {
-  backgroundColor: '#f9f9f9',
-  padding: '48px 80px',
-};
+const HRStyles = styled.hr`
+  border: 0;
+  margin-top: 0;
+  border-top: 1px solid ${(props) => props.theme.colors.grey_30};
+  margin-bottom: 0;
+`;
 
 class IndexPage extends React.Component {
   componentDidMount() {
-    const pix = document.createElement('script');
-    pix.language = 'JavaScript1.1';
-    pix.src = '//pixel.mathtag.com/event/js?mt_id=1538259&mt_adid=244742&mt_exem=&mt_excl=&v1=&v2=&v3=&s1=&s2=&s3=';
-    pix.async = true;
-    document.body.appendChild(pix);
+    const id = 'Polyfill';
+    if (!document.getElementById(id)) {
+      const polyfill = document.createElement('script');
+      polyfill.id = id;
+      polyfill.language = 'JavaScript1.1';
+      polyfill.src = '//polyfill.io/v3/polyfill.min.js?features=default%2CArray.prototype.find%2CArray.prototype.includes%2CPromise%2CObject.assign%2CObject.entries';
+      polyfill.async = true;
+      document.body.appendChild(polyfill);
+    }
   }
 
   render() {
-    const months = [
-      'Jan',
-      'Feb',
-      'Mar',
-      'Apr',
-      'May',
-      'Jun',
-      'Jul',
-      'Aug',
-      'Sep',
-      'Oct',
-      'Nov',
-      'Dec',
-    ];
-    const sortedUpcomingEvents = upcomingEvents;
-
     return (
       <Layout>
-        <SEO title="Home" slug="/" />
+        <SEO title="Learning Center" slug="/" />
         <div className="container-fluid">
-          <section className="row section align-items-center hero" style={heroBackground}>
+          <HeroWrapper className="row section align-items-center hero" >
             <div className="container">
               <div className="row">
-                <div className="col-sm-12 col-md-5 col-lg-6 align-self-center">
+                <div className="col-sm-12 col-md-5 col-lg-8 align-self-center">
                   <h1>Postman Learning Center</h1>
                   <p className="subtitle">
                     Learn how to use Postman.
@@ -71,13 +59,19 @@ class IndexPage extends React.Component {
                     <br />
                     Check out the docs and support resources!
                   </p>
-                  <a href="/docs/getting-started/introduction/" className="btn btn__primary-hollow mb-5">
-                    Explore the docs
-                  </a>
+                  <BaseButton
+                    src="/docs/introduction/overview/"
+                    className="mb-5 mb-md-0"
+                    as='a'
+                    buttonType="secondary"
+                    target="same-tab"
+                    >
+                    Explore the Docs
+                  </BaseButton>
                 </div>
-                <div className="col-sm-12 col-md-6 col-lg-6 align-self-center">
+                <div className="col-sm-12 col-md-6 col-lg-4 align-self-center">
                   <img
-                    src={HeroImage}
+                    src="https://voyager.postman.com/illustration/postman-learning-center-documentation-illustration.svg"
                     width="637"
                     height="411"
                     className="hero-image img-fluid"
@@ -86,75 +80,121 @@ class IndexPage extends React.Component {
                 </div>
               </div>
             </div>
-          </section>
+          </HeroWrapper>
         </div>
         <div className="container">
-          <section className="row section">
+          <SectionStyles className="row">
             <div className="col-sm-12">
               <h2 className="mb-5">Design, develop, and collaborate on your API projects</h2>
               <div className="row justify-content-center">
                 <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
                   <LandingCard
-                    title="Get Started"
+                    title="Get started with Postman"
                     description="Send your first API request in Postman in just a few clicks!"
                     cta="Send a request"
-                    link="/docs/getting-started/sending-the-first-request/"
-                    icon={LearnImage}
+                    link="/docs/getting-started/first-steps/sending-the-first-request/"
+                    icon="https://voyager.postman.com/icon/spaceship-rocket-launch-icon-postman.svg"
                   />
                 </div>
                 <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
                   <LandingCard
-                    title="Test"
-                    description="Write test scripts and build automation into your workflow."
-                    link="/docs/writing-scripts/intro-to-scripts/"
-                    cta="Create tests"
-                    icon={TestImage}
+                    title="Send requests"
+                    description="Send requests in Postman to connect to APIs you are working with."
+                    link="/docs/sending-requests/requests/"
+                    cta="Build API requests"
+                    icon="https://voyager.postman.com/icon/marketing-radar-icon-postman.svg"
                   />
                 </div>
                 <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                  <LandingCard
+                    title="API-first development"
+                    description="Use the API Builder to design your API in Postman."
+                    cta="Develop your API"
+                    link="/docs/designing-and-developing-your-api/the-api-workflow/"
+                    icon="https://voyager.postman.com/icon/api-first-company-postman-icon.svg"
+                  />
+                </div>
+                <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                  <LandingCard
+                    title="Test with Postman"
+                    description="Write test scripts and build automation into your workflow."
+                    link="/docs/writing-scripts/intro-to-scripts/"
+                    cta="Create tests"
+                    icon="https://voyager.postman.com/icon/flask-science-beaker-test-icon-postman.svg"
+                  />
+                </div>
+                <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                  <LandingCard
+                    title="Collaborate with your team"
+                    description="Use Postman to enhance collaboration within your team."
+                    link="/docs/collaborating-in-postman/collaborate-in-postman-overview/"
+                    cta="Start collaborating"
+                    icon="https://voyager.postman.com/icon/community-three-people-icon-postman.svg"
+                  />
+                </div>
+                <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                  <LandingCard
+                    title="Postman Flows"
+                    description="Use Postman's visual tool for creating API workflows."
+                    link="/docs/postman-flows/gs/flows-overview/"
+                    cta="Use Flows"
+                    icon="https://voyager.postman.com/icon/api-demo-icon-postman.svg"
+                  />
+                </div>
+              </div>
+            </div>
+          </SectionStyles>
+          <div className="container-fluid" >
+            <HRStyles/>
+          </div>
+          <SectionStyles className="row">
+            <div className="col-sm-12">
+              <h2 className="mb-5">Explore other Postman resources</h2>
+              <div className="row justify-content-center">
+                <div className="col-sm-6 mb-3 mb-md-4 pr-md-5">
                   <LandingCard
                     title="30 Days of Postman"
                     description="Tackle a new challenge each day with these developer tutorials."
                     link="https://www.postman.com/postman/workspace/30-days-of-postman-for-developers/overview"
                     cta="Start challenge"
-                    icon={ThirtyDaysImage}
+                    icon="https://voyager.postman.com/icon/trophy-award-icon-postman.svg"
                   />
                 </div>
-                <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                <div className="col-sm-6 mb-3 mb-md-4 pr-md-5">
                   <LandingCard
-                    title="Postman Space Camp"
-                    description="See previous and upcoming educational webinars."
-                    cta="See webinars"
-                    link="https://www.postman.com/events/postman-space-camp/"
-                    icon={SpaceCampImage}
+                    title="Postman Intergalactic"
+                    description="A series of educational trainings taught by Postman team members with a live Q&A."
+                    cta="See upcoming webinars"
+                    link="https://www.postman.com/events/intergalactic/"
+                    icon="https://voyager.postman.com/icon/product-ufo-icon-postman.svg"
                   />
                 </div>
-                <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                <div className="col-sm-6 mb-3 mb-md-4 pr-md-5">
                   <LandingCard
                     title="Postman Network"
                     description="Browse APIs, workspaces, and collections inside Postman."
                     link="https://www.postman.com/explore"
                     cta="Explore Postman"
-                    icon={NetworkImage}
+                    icon="https://voyager.postman.com/icon/hotspot-network-icon-postman.svg"
                   />
                 </div>
-                <div className="col-sm-6 col-lg-4 mb-3 mb-md-4 pr-md-5">
+                <div className="col-sm-6 mb-3 mb-md-4 pr-md-5">
                   <LandingCard
-                    title="Videos"
+                    title="Postman videos"
                     description="Learn Postman skills from our video playlists."
                     link="https://www.youtube.com/c/Postman"
                     cta="Watch videos"
-                    icon={VideoImage}
+                    icon="https://voyager.postman.com/icon/play-movie-video-film-icon-postman.svg"
                   />
                 </div>
               </div>
             </div>
-          </section>
+          </SectionStyles>
           <div className="container-fluid" >
-            <hr/>
+            <HRStyles/>
           </div>
           {/* Youtube Video Section */}
-          <section className="row section align-items-center">
+          <SectionStyles className="row align-items-center">
             <div className="col-lg-4">
               <h2>Intro to Postman</h2>
               <p>Learn the Postman fundamentals in this video course for beginners.</p>
@@ -163,137 +203,73 @@ class IndexPage extends React.Component {
               </p>
             </div>
             <div className="col-lg-8">
-              <div className="embed-responsive embed-responsive-16by9 img-frame">
-                <iframe
-                  loading="lazy"
-                  className="embed-responsive-item "
-                  src="https://www.youtube-nocookie.com/embed/2oOSnxZ28fA?rel=0"
-                  title="Intro to Postman"
-                  allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
-                  allowFullScreen
-                />
-              </div>
+              <VideoComponent
+                src="https://www.youtube.com/embed/2oOSnxZ28fA?rel=0&origin=https://learning.postman.com/"
+                border="true"
+              ></VideoComponent>
             </div>
-          </section>
+          </SectionStyles>
           <div className="container-fluid" >
-            <hr/>
+            <HRStyles/>
           </div>
           {/* Events Section */}
-          <section className="row section">
-            <div className="col-12 col col-lg-4 text-left mb-5">
-              <div className="sticky-top" style={{ top: '75px', zIndex: '0' }}>
-                <h2 id="upcoming-events">Upcoming Postman Events</h2>
-                <p>
-                  <a 
-                    className="link-style" 
-                    href="https://www.twitch.tv/getpostman" 
-                    target="_blank" 
-                    rel="noopener">
+          <SectionStyles className="row align-items-center hero" >
+            <div className="container">
+              <div className="row">
+                <div className="col-sm-12 col-md-5 col-lg-8 align-self-center">
+                  <h2 id="upcoming-events">Upcoming Postman events</h2>
+                  <p>
+                  <BaseLink
+                    src="https://www.twitch.tv/getpostman"
+                    target="new-tab-external-nofollow"
+                    >
                     Follow us
-                  </a>
+                  </BaseLink>
                   {' '}
                   on Twitch or
                   {' '}
-                  <a
-                    className="link-style"
-                    href="https://www.youtube.com/channel/UCocudCGVb3MmhWQ1aoIgUQw"
-                    target="_blank"
-                    rel="noopener"
+                  <BaseLink
+                    src="https://www.youtube.com/channel/UCocudCGVb3MmhWQ1aoIgUQw"
+                    target="new-tab-external-nofollow"
                   >
                     subscribe
-                  </a>
+                  </BaseLink>
                   {' '}
                   to our YouTube channel so you don’t miss when we go live.
                 </p>
+                  <BaseButton
+                    src="https://www.postman.com/events/"
+                    className="mb-5 mb-md-0"
+                    as='a'
+                    buttonType="secondary"
+                    target="same-tab"
+                    >
+                    View Upcoming Events
+                  </BaseButton>
+                </div>
+                <div className="col-sm-12 col-md-6 col-lg-4 align-self-center">
+                  <img
+                    src="https://voyager.postman.com/illustration/webinar-illustration-postman.svg"
+                    width="637"
+                    height="411"
+                    className="hero-image img-fluid"
+                    alt="Postmanaut sitting at computer. Illustration."
+                  />
+                </div>
               </div>
             </div>
-            <div className="col-12 col-lg-8 justify-content-center d-flex align-items-stretch event-wrapper">
-              <div className="row">
-                {/* If there are events in the events.json array */}
-                {sortedUpcomingEvents.length > 0 ? (
-                  // Map over, get the appropriate values and render event.
-                  sortedUpcomingEvents.map((event) => {
-                    // Used for Left side of events (Example: JAN 1)
-                    const dateObject = new Date(event.date);
-                    // Used for Right Side of events
-                    const today = new Date(event.date);
-                    const date = `${today.getMonth() + 1}/${today.getDate()}/${today.getFullYear()}`;
-                    // Combine platform, date and time of the event
-                    // Example: Livestream - 01/01/2000 4PM PST
-                    const eventInformation = `${event.location} - ${date} ${event.time}`;
-                    // added parathensis to remove linter error
-                    const isStale = new Date(event.expirationDate) < ((new Date()) && ' d-none') || '';
-                    return (
-                      <div className={`col-12 col-lg-10 offset-lg-1 ${isStale} event-single-wrapper`} key={uuidv4()}>
-                        <div className="row">
-                          <div className="col-12 col-lg-3 event-date event-month">
-                            {/* <span className="event-month"> */}
-                            {`${months[dateObject.getUTCMonth()]}`}
-                            {/* </span> */}
-                            {' '}
-                            {`${dateObject.getDate()}`}
-                          </div>
-                          <div className="col-12 col-lg-9 event-description-wrapper">
-                            <p className="mb-1 event-location">{`${eventInformation}`}</p>
-                            <h4 className="event-title">{event.title}</h4>
-                            <p>{event.description}</p>
-                            <OutboundLink
-                              className="event-link-wrapper link-style"
-                              href={event.link}
-                              target="_blank"
-                              rel="noopener"
-                            >
-                              <span>See details →</span>
-                            </OutboundLink>
-                          </div>
-                        </div>
-                      </div>
-                    );
-                  })
-                ) : (
-                  <>
-                    {/* If there are no events, and events.json is an object
-                      where development eq true */}
-                    {!Array.isArray(upcomingEvents) && upcomingEvents.development ? (
-                      <div className="events__alert" role="alert">
-                        <p>
-                          You are currently in develop mode. Dynamic events will not be displayed
-                          locally.
-                          <a
-                            className="link-style"
-                            style={{ fontSize: 'inherit' }}
-                            href="https://github.com/postmanlabs/postman-docs/blob/develop/CONTRIBUTING.md"
-                            target="_blank"
-                            rel="noopener"
-                          >
-                            See Contributing doc for details
-                          </a>
-                          .
-                        </p>
-                      </div>
-                    ) : (
-                      <>
-                        {/* else we know we have 0 upcoming events, and we are not
-                          in development mode */}
-                        <p>We currently have no upcoming events...check back later.</p>
-                      </>
-                    )}
-                  </>
-                )}
-              </div>
-            </div>
-          </section>
+          </SectionStyles>
           <div className="container-fluid" >
-            <hr/>
+            <HRStyles/>
           </div>
-          <section className="row section">
+          <SectionStyles className="row">
             <div className="col-sm-6 col-lg-3 mb-sm-4 mb-md-0 pr-md-5">
               <LandingCard
                 title="Postman support"
-                description="Reach out to our support team."
-                cta="Open a ticket"
-                icon={SupportImage}
-                link="https://support.postman.com/hc/en-us/requests/new/"
+                description="Get help for your issue or a specific question."
+                cta="Visit Postman support center"
+                icon="https://voyager.postman.com/icon/support-life-ring-icon-postman.svg"
+                link="https://support.postman.com/hc/en-us"
               />
             </div>
             <div className="col-sm-6 col-lg-3 mb-sm-4 mb-md-0 pr-md-5">
@@ -301,29 +277,29 @@ class IndexPage extends React.Component {
                 title="Bugs and feature requests"
                 description="Check out the app support repo."
                 cta="Make a request"
-                icon={BugImage}
+                icon="https://voyager.postman.com/icon/bug-error-icon-postman.svg"
                 link="https://github.com/postmanlabs/postman-app-support/"
               />
             </div>
             <div className="col-sm-6 col-lg-3 mb-sm-4 mb-md-0 pr-md-5">
               <LandingCard
-                title="Community"
+                title="Postman community"
                 description="Join the Postman community."
                 cta="Visit forum"
-                icon={CommunityImage}
+                icon="https://voyager.postman.com/icon/community-three-people-icon-postman.svg"
                 link="https://community.postman.com/"
               />
             </div>
             <div className="col-sm-6 col-lg-3 mb-sm-4 mb-md-0">
               <LandingCard
-                title="Postman Answers"
+                title="Postman answers"
                 description="Code samples for most commonly asked questions."
-                cta="Visit Postman Answers"
-                icon={SolutionImage}
+                cta="Visit Postman answers"
+                icon="https://voyager.postman.com/icon/solution-puzzle-answers-icon-postman.svg"
                 link="https://www.postman.com/postman/workspace/postman-answers/"
               />
             </div>
-          </section>
+          </SectionStyles>
         </div>
       </Layout>
     );
